4.4.18 changes and thoughts
there was a probloem with the way we thought last time we cant use the 5 lines hourglass because the right-side spacings. so in order to achieve the necessery functionality i chnaged the beginner hourglass function to create a sand clock by parameter and we'll refrence this function in advanced hourglass whos file isnt finished!.
the testing frame work is built and working ill add the files .
tested on the hour glass for inputs 0-100  also test files well be submitted.